Began improving and refactoring RoleManager. Fixed some bugs with static factory and modules
