+++ controllers/roles/_manage_role.htm
@@ -19,44 +19,47 @@
 
 
     ?>
+    <div class="table-responsive">
 
-    <table id="role_table" style="table-layout:fixed; width:800px;">
+        <table id="role_table" class="table table-condensed table-hover table-striped" style="table-layout:fixed;">
 
-        <?php if(count($role->getUsersInRole()) != 0): ?>
+            <?php if(count($role->getUsersInRole()) != 0): ?>
 
-        <tr>
-            <th style="width:25px;"></th>
-            <th style="width:25px;"></th>
-            <th style="width:25px;"></th>
-            <th style="width:100px;">Name</th>
-            <th style="padding-left:25px;width:250px;">Email</th>
-            <th style="padding-left:25px;width:150px;">Created</th>
-            <th style="padding-left:25px;width:150px;">Last Seen</th>
-        </tr>
+            <tr>
+                <th style="width:25px;"></th>
+                <th style="width:25px;"></th>
+                <th style="width:25px;"></th>
+                <th style="width:125px;">Name</th>
+                <th style="padding-left:25px;">Email</th>
+                <th style="padding-left:25px;">Created</th>
+                <th style="padding-left:25px;">Last Seen</th>
+            </tr>
 
-        <?php foreach($role->getUsersInRole() as $index => $user): ?>
-        <tr style="height:25px;">
-            <div style="margin-top:9px;">
-                <td><span><a style="" href="#" data-request-confirm="Are you sure you want to remove this user from this role?" data-request="onRemoveUserFromRole" data-request-data="roleId: '<?= $role->id ?>', userId: '<?= $user->id ?>'"><i class="icon-times"></i></a></span></td>
-                <td><span><a style=" <?php if($role->sort_order == 1): ?> pointer-events: none; cursor: default; color:#BABABA; <?php endif ?> " href="#" data-request-confirm="Are you sure you want to promote this user to the next role?" data-request="onPromote" data-request-data="roleCode: '<?= $role->code ?>', userId: '<?= $user->id ?>'"><i class="icon-caret-up"></i></a></span></td>
-                <td><span><a style=" <?php if($role->sort_order == $roleTotal): ?> pointer-events: none; cursor: default; color:#BABABA; <?php endif ?> " href="#" data-request-confirm="Are you sure you want to demote this user down a role?" data-request="onDemote" data-request-data="roleCode: '<?= $role->code ?>', userId: '<?= $user->id ?>'"><i class="icon-caret-down"></i></a></span></td>
-                <td style="width:150px;overflow:hidden;white-space: nowrap;"><span><a href="<?= Backend::url('/rainlab/user/users/preview/'.$user->id) ?>"><?= $user->name ?></a></span></td>
-                <td style="padding-left:25px;width:250px;overflow:hidden;white-space: nowrap;"><span><?= $user->email ?></span></td>
-                <td style="padding-left:25px;width:150px;overflow:hidden;white-space: nowrap;"><span><?= $user->created_at ?></span></td>
-                <td style="padding-left:25px;width:150px;overflow:hidden;white-space: nowrap;"><span><?= $user->last_seen ?></span></td>
-            </div>
-        </tr>
-        <?php endforeach; else: ?>
-        <tr>
-            <td>
-                <span></span>
-            </td>
-        </tr>
-        <?php endif ?>
+            <?php foreach($role->getUsersInRole() as $index => $user): ?>
+            <tr style="height:25px;">
+                <div style="margin-top:9px;">
+                    <td><span><a style="" href="#" data-request-confirm="Are you sure you want to remove this user from this role?" data-request="onRemoveUserFromRole" data-request-data="roleId: '<?= $role->id ?>', userId: '<?= $user->id ?>'"><i class="icon-times"></i></a></span></td>
+                    <td><span><a style=" <?php if($role->sort_order == 1): ?> pointer-events: none; cursor: default; color:#BABABA; <?php endif ?> " href="#" data-request-confirm="Are you sure you want to promote this user to the next role?" data-request="onPromote" data-request-data="roleCode: '<?= $role->code ?>', userId: '<?= $user->id ?>'"><i class="icon-caret-up"></i></a></span></td>
+                    <td><span><a style=" <?php if($role->sort_order == $roleTotal): ?> pointer-events: none; cursor: default; color:#BABABA; <?php endif ?> " href="#" data-request-confirm="Are you sure you want to demote this user down a role?" data-request="onDemote" data-request-data="roleCode: '<?= $role->code ?>', userId: '<?= $user->id ?>'"><i class="icon-caret-down"></i></a></span></td>
+                    <td style="overflow:hidden;white-space: nowrap;"><span><a href="<?= Backend::url('/rainlab/user/users/preview/'.$user->id) ?>"><?= $user->name ?></a></span></td>
+                    <td style="padding-left:25px;overflow:hidden;white-space: nowrap;"><span><?= $user->email ?></span></td>
+                    <td style="padding-left:25px;overflow:hidden;white-space: nowrap;"><span><?= $user->created_at ?></span></td>
+                    <td style="padding-left:25px;overflow:hidden;white-space: nowrap;"><span><?= $user->last_seen ?></span></td>
+                </div>
+            </tr>
+            <?php endforeach; else: ?>
+            <tr>
+                <td>
+                    <span></span>
+                </td>
+            </tr>
+            <?php endif ?>
 
 
 
-    </table>
+        </table>
+
+    </div>
 
 
 </div>
