+++ controllers/roles/_management_role_toolbar.htm
@@ -6,7 +6,7 @@
             <span style="font-size:19px;font-style:italic;">Selected Role: <?= $role->name ?></span>
         </div>
 
-        <div class="col-sm-12 col-md-6">
+        <div class="col-sm-12 col-md-6 text-right">
             <!-- <a
                      href="javascript:;"
                      class="btn btn-primary oc-icon-plus"
